+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## New Features
 
 - **Implemented experimental 4X (800p) and 8X (1600p) resolutions**
+- **Implemented chasecam**
 - **Arms number 1 can now be colored red if Berserk is available**
 
 ## Changes